+++ src/server/controllers/continue.controller.js
@@ -33,21 +33,27 @@ const continueController = (
       statusEmitter.emit("incrementCount", "registrationsStarted");
     }
 
-    const newAnswersArray = Object.values(newAnswers);
+    const trimmedNewAnswers = JSON.parse(JSON.stringify(newAnswers));
+
+    for (let answer in trimmedNewAnswers) {
+      trimmedNewAnswers[answer] = trimmedNewAnswers[answer].trim();
+    }
+
+    const trimmedNewAnswersArray = Object.values(trimmedNewAnswers);
 
     let cleanedPreviousAnswers = Object.assign({}, previousAnswers);
 
     // remove any answers that were previously given a truthy value but have since been emptied
     cleanedPreviousAnswers = cleanEmptiedAnswers(
       previousAnswers,
-      newAnswersArray,
+      trimmedNewAnswersArray,
       currentPage
     );
 
     controllerResponse.cumulativeAnswers = Object.assign(
       {},
       cleanedPreviousAnswers,
-      newAnswers
+      trimmedNewAnswers
     );
 
     controllerResponse.switches = Object.assign(
@@ -57,7 +63,7 @@ const continueController = (
 
     controllerResponse.validatorErrors = Object.assign(
       {},
-      validate(currentPage, newAnswers).errors
+      validate(currentPage, trimmedNewAnswers).errors
     );
 
     if (Object.keys(controllerResponse.validatorErrors).length > 0) {